adding 3 files
One JSON file for each screen of the Translator App, prepended with the screen name.
One XML file for each set of screen blocks (the second screen in the example has not blocks, so we only see one XML file here),  prepended with the screen name.
One meta.json file used to describe the entire app